Fixes #62.
* Renamed all `.h` headers to `.hpp`
* Made the examples take our headers from the system-path (i.e. `<>` over `""` in inclusion)
