+++ examples/by_runtime_api_module/error_handling.cu
@@ -5,7 +5,7 @@
  *   Error Handling
  *
  */
-#include <cuda/api_wrappers.h>
+#include <cuda/api_wrappers.hpp>
 
 #include <iostream>
 #include <string>
