+++ examples/by_runtime_api_module/execution_control.cu
@@ -12,7 +12,7 @@
  *   cudaSetDoubleForHost
  *
  */
-#include "cuda/api_wrappers.h"
+#include <cuda/api_wrappers.hpp>
 
 #include <cuda_runtime_api.h>
 #if __CUDACC_VER_MAJOR__ >= 9
