+++ examples/by_runtime_api_module/stream_management.cu
@@ -5,7 +5,7 @@
  *   Stream Management
  *
  */
-#include "cuda/api_wrappers.h"
+#include <cuda/api_wrappers.hpp>
 
 #include <cuda_runtime_api.h>
