+++ examples/modified_cuda_samples/helper_cuda.h
@@ -22,7 +22,7 @@
 #include <stdio.h>
 #include <string.h>
 
-#include "cuda/api_wrappers.h"
+#include <cuda/api_wrappers.hpp>
 
 #include <iostream>
 #include <iomanip>
