+++ examples/modified_cuda_samples/simpleIPC/simpleIPC.cu
@@ -17,7 +17,7 @@
 // CUDA runtime includes
 #include <cuda_runtime_api.h>
 
-#include "cuda/api_wrappers.h"
+#include <cuda/api_wrappers.hpp>
 
 #include "../helper_string.h"
