+++ examples/other/io_compute_overlap_with_streams.cu
@@ -5,7 +5,7 @@
  *   Stream Management
  *
  */
-#include "cuda/api_wrappers.h"
+#include <cuda/api_wrappers.hpp>
 
 #include <iostream>
 #include <vector>
