+++ src/cuda/api/constants.hpp
@@ -10,7 +10,7 @@
 #ifndef CUDA_API_WRAPPERS_CONSTANTS_H_
 #define CUDA_API_WRAPPERS_CONSTANTS_H_
 
-#include <cuda/api/types.h>
+#include <cuda/api/types.hpp>
 
 namespace cuda {
