+++ src/cuda/api/device.hpp
@@ -9,11 +9,11 @@
 #ifndef CUDA_API_WRAPPERS_DEVICE_HPP_
 #define CUDA_API_WRAPPERS_DEVICE_HPP_
 
-#include <cuda/api/types.h>
+#include <cuda/api/types.hpp>
 #include <cuda/api/device_properties.hpp>
 #include <cuda/api/memory.hpp>
 #include <cuda/api/current_device.hpp>
-#include <cuda/api/pci_id.h>
+#include <cuda/api/pci_id.hpp>
 #include <cuda/api/unique_ptr.hpp>
 
 #include <cuda_runtime_api.h>
