+++ src/cuda/api/device_count.hpp
@@ -13,7 +13,7 @@
 #ifndef CUDA_API_WRAPPERS_DEVICE_COUNT_HPP_
 #define CUDA_API_WRAPPERS_DEVICE_COUNT_HPP_
 
-#include <cuda/api/types.h>
+#include <cuda/api/types.hpp>
 #include <cuda_runtime_api.h>
 
 namespace cuda {
