+++ src/cuda/api/device_function.hpp
@@ -13,7 +13,7 @@
 #ifndef CUDA_API_WRAPPERS_DEVICE_FUNCTION_HPP_
 #define CUDA_API_WRAPPERS_DEVICE_FUNCTION_HPP_
 
-#include <cuda/api/types.h>
+#include <cuda/api/types.hpp>
 #include <cuda/api/device_properties.hpp>
 #include <cuda/api/error.hpp>
 #include <cuda/api/current_device.hpp>
