+++ src/cuda/api/error.hpp
@@ -9,7 +9,7 @@
 #ifndef CUDA_API_WRAPPERS_ERROR_HPP_
 #define CUDA_API_WRAPPERS_ERROR_HPP_
 
-#include <cuda/api/types.h>
+#include <cuda/api/types.hpp>
 
 #include <cuda_runtime_api.h>
 #include <type_traits>
