+++ src/cuda/api/kernel_launch.cuh
@@ -42,8 +42,8 @@
 #ifndef CUDA_API_WRAPPERS_KERNEL_LAUNCH_CUH_
 #define CUDA_API_WRAPPERS_KERNEL_LAUNCH_CUH_
 
-#include <cuda/api/types.h>
-#include <cuda/api/constants.h>
+#include <cuda/api/types.hpp>
+#include <cuda/api/constants.hpp>
 #include <cuda/api/device_function.hpp>
 
 #if (__CUDACC_VER_MAJOR__ >= 9)
