+++ src/cuda/api/memory.hpp
@@ -9,7 +9,7 @@
 #define CUDA_API_WRAPPERS_MEMORY_HPP_
 
 #include <cuda/api/error.hpp>
-#include <cuda/api/constants.h>
+#include <cuda/api/constants.hpp>
 #include <cuda/api/current_device.hpp>
 
 #include <cuda_runtime.h> // needed, rather than cuda_runtime_api.h, e.g. for cudaMalloc
