+++ src/cuda/api/pci_id.hpp
@@ -1,68 +1,57 @@
 /**
- * @file pci_id.hpp
+ * @file pci_id.h
  *
- * @brief iostream-related freestanding operator functions for
- * @ref cuda::device::pci_location_t instances and iostream-related methods of
- * the @ref cuda::device::pci_location_t class.
- *
- * @note This file is split off from {@ref pci_id.h} since
- * it requires inclusions of standard library headers which most of the
- * API wrapper code doesn't.
+ * @brief Definition of a wrapper class for CUDA PCI device ID
+ * information.
  */
 #pragma once
-#ifndef CUDA_API_WRAPPERS_PCI_ID_HPP_
-#define CUDA_API_WRAPPERS_PCI_ID_HPP_
+#ifndef CUDA_API_WRAPPERS_PCI_ID_H_
+#define CUDA_API_WRAPPERS_PCI_ID_H_
 
-#include <cuda/api/pci_id.h>
+#include <cuda/api/types.hpp>
+#include <cuda/api/error.hpp>
+
+#include <cuda_runtime_api.h>
 
 #include <string>
-#include <istream>
-#include <ostream>
-#include <sstream>
-
 
 namespace cuda {
 namespace device {
 
-///@cond
+/**
+ * Location "coordinates" for a CUDA device on a PCIe bus
+ *
+ * @note can be compiled from individual values from a device's properties;
+ * see @ref properties_t
+ */
+struct pci_location_t {
+	// This is simply what we get in CUDA's cudaDeviceProp structure
+	int domain;
+	int bus;
+	int device;
 
-inline std::istream& operator>>(std::istream& is, cuda::device::pci_location_t& pci_id)
+	operator std::string() const;
+	// This is not a ctor so as to maintain the PODness
+	static pci_location_t parse(const std::string& id_str);
+};
+
+/**
+ * Obtain a CUDA device id for a PCIe bus device
+ *
+ * @param pci_id the location on (one of) the PCI bus(es) of
+ * the device of interest
+ */
+inline id_t resolve_id(pci_location_t pci_id)
 {
-	auto format_flags(is.flags());
-	is >> std::hex;
-	is >> pci_id.domain; is.ignore(1); // ignoring a ':'
-	is >> pci_id.bus;    is.ignore(1); // ignoring a ':'
-	is >> pci_id.device;
-	is.flags(format_flags);
-	return is;
+	std::string as_string = pci_id;
+	id_t cuda_device_id;
+	auto result = cudaDeviceGetByPCIBusId(&cuda_device_id, as_string.c_str());
+	throw_if_error(result,
+		"Failed obtaining a CUDA device ID corresponding to PCI id " + as_string);
+	return cuda_device_id;
 }
 
-inline std::ostream& operator<<(std::ostream& os, const cuda::device::pci_location_t& pci_id)
-{
-	auto format_flags(os.flags());
-	os << std::hex << pci_id.domain << ':' << pci_id.bus << ':' << pci_id.device;
-	os.flags(format_flags);
-	return os;
-}
-
-inline pci_location_t::operator std::string() const
-{
-	std::ostringstream oss;
-	oss << (*this);
-	return oss.str();
-}
-
-inline pci_location_t pci_location_t::parse(const std::string& id_str)
-{
-	std::istringstream iss(id_str);
-	pci_location_t id;
-	iss >> id;
-	return id;
-}
-
-///@endcond
-
-} //namespace device
+} // namespace device
 } // namespace cuda
 
-#endif /* CUDA_API_WRAPPERS_PCI_ID_HPP_ */
+#endif /* CUDA_API_WRAPPERS_PCI_ID_H_ */
