+++ src/cuda/api_wrappers.hpp
@@ -8,15 +8,14 @@
 #ifndef CUDA_API_WRAPPERS_H_
 #define CUDA_API_WRAPPERS_H_
 
-#include <cuda/api/types.h>
-#include <cuda/api/constants.h>
+#include <cuda/api/types.hpp>
+#include <cuda/api/constants.hpp>
 #include <cuda/api/error.hpp>
 #include <cuda/api/versions.hpp>
 #include <cuda/api/miscellany.hpp>
 #include <cuda/api/kernel_launch.cuh>
-#include <cuda/api/profiling.h>
+#include <cuda/api/profiling.hpp>
 #include <cuda/api/device_properties.hpp>
-#include <cuda/api/pci_id.hpp>
 #include <cuda/api/device_count.hpp>
 #include <cuda/api/current_device.hpp>
 #include <cuda/api/device_function.hpp>
@@ -29,6 +28,7 @@
 #include <cuda/api/device.hpp>
 #include <cuda/api/event.hpp>
 
+#include <cuda/api/pci_id_impl.hpp>
 #include <cuda/api/multi_wrapper_impls.hpp>
 
 #endif /* CUDA_API_WRAPPERS_H_ */
